Use Ionic.Zip for APK updating
On Android, System.IO.Compression likes to force compress the zip for some reason, which leads to Android complaining about some APKs not being valid. I have no idea why this happens but this seemed to fix it.

IO.Compression still works fine for extractions so I just left it